Improved user feedback for labs_SpawnFromTemplate
In the previous version the Action "Create New From Template" will throw the following error message when action is performed on an item that is no template: "The method "labs_SpawnFromTemplate" failed." This default message is not very helpful for the end user. This change adds a more user friendly text for the case users want to create new documents from a non-template document.